requirements.txt
pandas
numpy
scikit-learn
xgboost
mlflow
dvc
fastapi
uvicorn
pydantic
joblib
apache-airflow
pypdf
